+++ src/main/java/com/parkify/vehitrack/exception/InvalidRegistrationNumberException.java
@@ -1,0 +1,8 @@
+package com.parkify.vehitrack.exception;
+
+public class InvalidRegistrationNumberException extends RuntimeException{
+
+    public InvalidRegistrationNumberException(String message) {
+        super(message);
+    }
+}